azfunc-selenium-docker-sample/HttpExample/__init__.py
import logging
import azure.functions as func

from selenium import webdriver   # for webdriver
from selenium.webdriver.support.ui import WebDriverWait  # for implicit and explict waits
from selenium.webdriver.chrome.options import Options  # for suppressing the browser


def getWebSiteTitle(url):
    chrome_options = Options()
    chrome_options.add_argument('--headless')
    chrome_options.add_argument('--no-sandbox')
    chrome_options.add_argument('--disable-dev-shm-usage')
    # abosulte path to copied binary
    driver = webdriver.Chrome('/home/site/wwwroot/HttpExample/chromedriver',chrome_options=chrome_options)
    driver.get(url)
    return driver.title

def main(req: func.HttpRequest) -> func.HttpResponse:
    logging.info('Python HTTP trigger function processed a request.')

    url = req.params.get('url')
    if not url:
        try:
            req_body = req.get_json()
        except ValueError:
            pass
        else:
            url = req_body.get('url')

    if url:
        title = getWebSiteTitle(url)
        return func.HttpResponse(f"The title of {url} is: '{title}'. This HTTP triggered function executed successfully.")
    else:
        return func.HttpResponse(
             "This HTTP triggered function executed successfully. Pass a url in the query string or in the request body for a personalized response.",
             status_code=200
        )